Add files via upload
This program is used to calculate experimental branching fractions and transition probabilities from measurements of atomic spectra. The program is used with spectral line lists from intensity-calibrated spectra, such as those observed with a Fourier transform spectrometer. The program takes a set of linelists and computes a weighted average branching fraction for each line from an upper level to lower level, using the calibrated integrated intensities and signal-to-noise  ratios. It combines these branching fractions with the level lifetime to obtain the transition probability. 

The documentation is in the file pybranch.pdf.